move header's logic into Presence-header
