+++ src/components/deputy/presence/Presence.tsx
@@ -2,6 +2,8 @@ import React, { useState } from "react"
 import Block from "../_block/_Block"
 import PresenceParticipation from "src/components/charts/PresenceParticipation"
 import { ParentSize } from "@visx/responsive"
+import { orderBy } from "lodash"
+import PresenceHeader from "./Presence-header"
 
 /**
  * Return deputy's presence and participation graph in a Block component
@@ -9,17 +11,25 @@ import { ParentSize } from "@visx/responsive"
  */
 
 const Presence = (props: Bloc.Presence) => {
-  const [HasInformations, setHasInformations] = useState(false)
+  const orderedWeeks = orderBy(props.activite, "DateDeFin")
+  const [RangeOrderedWeeks, setRangeOrderedWeeks] = useState(orderedWeeks)
+
   return (
     <Block title="Présence et participation" type="presence" color={props.color} size={props.size}>
       <ParentSize debounceTime={400}>
         {(parent) => (
-          <PresenceParticipation
-            width={parent.width}
-            height={parent.height * 0.9}
-            data={props.activite}
-            color={props.color.HSL.Full}
-          />
+          <>
+            <div className="presence">
+              <PresenceHeader width={parent.width} data={orderedWeeks} setRange={setRangeOrderedWeeks} />
+              <PresenceParticipation
+                width={parent.width}
+                height={parent.height * 0.9}
+                data={orderedWeeks}
+                slicedData={RangeOrderedWeeks}
+                color={props.color.HSL.Full}
+              />
+            </div>
+          </>
         )}
       </ParentSize>
     </Block>
